// src/telegram/enum/tokens/telegram.tokens.enum.ts
export enum TelegramTokensEnum {
  TELEGRAM_SERVICE_TOKEN = 'TELEGRAM_SERVICE_TOKEN',
  TELEGRAM_UPDATE_TOKEN = 'TELEGRAM_UPDATE_TOKEN',
  TELEGRAM_MENU_WIZARD_TOKEN = 'TELEGRAM_MENU_WIZARD_TOKEN',
  TELEGRAM_MAKE_ADMIN_WIZARD_TOKEN = 'TELEGRAM_MAKE_ADMIN_WIZARD_TOKEN',
  TELEGRAM_MAKE_SELLER_WIZARD_TOKEN = 'TELEGRAM_MAKE_SELLER_WIZARD_TOKEN',
  TELEGRAM_ADMIN_MENU_WIZARD_TOKEN = 'TELEGRAM_ADMIN_MENU_WIZARD_TOKEN',
  TELEGRAM_SELLER_MENU_WIZARD_TOKEN = 'TELEGRAM_SELLER_MENU_WIZARD_TOKEN',
  TELEGRAM_SELLERS_WIZARD_TOKEN = 'TELEGRAM_SELLERS_WIZARD_TOKEN',
}