// src/telegram/enum/tokens/telegram.wizards.tokens.enum.ts
export enum TelegramWizardsTokensEnum {
  TELEGRAM_ORDER_CALL_WIZARD_TOKEN = 'TELEGRAM_ORDER_CALL_WIZARD_TOKEN',
  TELEGRAM_BOOK_EVENT_WIZARD_TOKEN = 'TELEGRAM_BOOK_EVENT_WIZARD_TOKEN',
  TELEGRAM_MAKE_ADMIN_WIZARD_TOKEN = 'TELEGRAM_MAKE_ADMIN_WIZARD_TOKEN',
  TELEGRAM_MAKE_SELLER_WIZARD_TOKEN = 'TELEGRAM_MAKE_SELLER_WIZARD_TOKEN',
  TELEGRAM_ADMIN_MENU_WIZARD_TOKEN = 'TELEGRAM_ADMIN_MENU_WIZARD_TOKEN',
  TELEGRAM_SELLER_MENU_WIZARD_TOKEN = 'TELEGRAM_SELLER_MENU_WIZARD_TOKEN',
  TELEGRAM_SELLERS_WIZARD_TOKEN = 'TELEGRAM_SELLERS_WIZARD_TOKEN',
}